Move HelpRequest classes to help package.
And delete arg_splitter.py.
